# Conflicts:
#	bin/ch/WScriptJsrt.cpp
#	lib/Backend/CaseNode.h
#	lib/Backend/Chakra.Backend.vcxproj
#	lib/Backend/Func.cpp
#	lib/Backend/IRBuilderAsmJs.cpp
#	lib/Backend/Lower.cpp
#	lib/Backend/LowerMDShared.cpp
#	lib/Backend/NativeCodeGenerator.cpp
#	lib/Backend/SwitchIRBuilder.cpp
#	lib/Parser/Chakra.Parser.vcxproj
#	lib/Runtime/Base/Chakra.Runtime.Base.vcxproj
#	lib/Runtime/Language/AsmJsTypes.h
#	lib/Runtime/Language/Chakra.Runtime.Language.vcxproj.filters

+++ lib/Runtime/ByteCode/ByteCodeSerializer.h
@@ -84,10 +84,10 @@ namespace Js
         int magic; // magicStartOfAuxPropIdArray
 #endif
         int propertyCount;
-        int extraSlots;
+        byte extraSlots;
         bool hadDuplicates;
         bool has__proto__;
-        SerializedPropertyIdArray(uint offset, int propertyCount, int extraSlots, bool hadDuplicates, bool has__proto__);
+        SerializedPropertyIdArray(uint offset, int propertyCount, byte extraSlots, bool hadDuplicates, bool has__proto__);
     };
 
     // The in-memory layout of the serialized analog of FuncInfoArray
